refact: Relocating QuantileDeltaMappingCorrection
The bias_calc was already quite large before adding PresRat. This commit
relocates QuantileDeltaMappingCorrection in a new sub-module that will
also receive PresRat.
