+++ sup3r/bias/__init__.py
@@ -3,5 +3,5 @@
 from .bias_transforms import (global_linear_bc, local_linear_bc,
                               monthly_local_linear_bc)
 from .bias_calc import (LinearCorrection, MonthlyLinearCorrection,
-                        QuantileDeltaMappingCorrection,
                         SkillAssessment)
+from .qdm import QuantileDeltaMappingCorrection
